pullup_bot/requirements.txt
discord.py
mediapipe
opencv-python
aiohttp
